requirements.txt
Flask==1.0.2
nltk==3.4.5
colorama==0.4.1
rake_nltk==1.0.4
wikipedia==1.4.0
apscheduler==3.8.1